Moved NullableConversion method to ParameterConversions to minimize file diffs. Technicall nothing chnages. The method still needs access to ParameterConversionConfiguration.ParameterConversions
